Improvements to JPA annotations (e.g. length values set where appropriate) and raw data model start / stop coordinates changed to int values to be constistent with the post-processed model.
git-svn-id: https://interproscan.googlecode.com/svn/trunk@326 ea5da8b0-6d6b-11de-9de6-4d42d2ab87c0

+++ core/model/src/main/java/uk/ac/ebi/interpro/scan/model/raw/Gene3dHmmer3RawMatch.java
@@ -1,6 +1,7 @@
 package uk.ac.ebi.interpro.scan.model.raw;
 
 import javax.persistence.Entity;
+import javax.persistence.Column;
 
 /**
  * TODO: Add class description
@@ -18,9 +19,9 @@ public class Gene3dHmmer3RawMatch extends Hmmer3RawMatch {
 
     public Gene3dHmmer3RawMatch(String sequenceIdentifier, String model,
                                 String signatureLibraryName, String signatureLibraryRelease,
-                                long locationStart, long locationEnd,
+                                int locationStart, int locationEnd,
                                 double evalue, double score,
-                                long hmmStart, long hmmEnd, String hmmBounds,
+                                int hmmStart, int hmmEnd, String hmmBounds,
                                 double locationScore,
                                 int envelopeStart, int envelopeEnd,
                                 double expectedAccuracy, double fullSequenceBias,
