+++ core/model/src/main/java/uk/ac/ebi/interpro/scan/model/raw/ProSiteProfileRawMatch.java
@@ -15,7 +15,7 @@ public class ProSiteProfileRawMatch extends ProfileScanRawMatch {
 
     public ProSiteProfileRawMatch(String sequenceIdentifier, String model,
                                   String signatureLibraryName, String signatureLibraryRelease,
-                                  long locationStart, long locationEnd,
+                                  int locationStart, int locationEnd,
                                   double score, String generator) {
         super(sequenceIdentifier, model, signatureLibraryName, signatureLibraryRelease, locationStart, locationEnd, score, generator);
     }
